verbose: render dynamic match locations
